# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ChannelClockSummary": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "confidence",
          "summary"
        ],
        "title": "ChannelClockSummary",
        "type": "object"
      },
      "ChannelTimeline": {
        "additionalProperties": false,
        "properties": {
          "channel": {
            "title": "Channel",
            "type": "integer"
          },
          "clock": {
            "$ref": "#/components/schemas/ChannelClockSummary"
          },
          "color_token": {
            "title": "Color Token",
            "type": "string"
          },
          "coverage": {
            "items": {
              "maxItems": 3,
              "minItems": 3,
              "prefixItems": [
                {
                  "type": "integer"
                },
                {
                  "type": "integer"
                },
                {
                  "enum": [
                    "index",
                    "carved",
                    "inferred"
                  ],
                  "type": "string"
                }
              ],
              "type": "array"
            },
            "title": "Coverage",
            "type": "array"
          },
          "deleted": {
            "items": {
              "maxItems": 3,
              "minItems": 3,
              "prefixItems": [
                {
                  "type": "integer"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "string"
                }
              ],
              "type": "array"
            },
            "title": "Deleted",
            "type": "array"
          },
          "image_id": {
            "title": "Image Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "motion": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "title": "Motion",
            "type": "array"
          }
        },
        "required": [
          "image_id",
          "channel",
          "label",
          "color_token",
          "coverage",
          "deleted",
          "motion",
          "clock"
        ],
        "title": "ChannelTimeline",
        "type": "object"
      },
      "ClockModel": {
        "additionalProperties": false,
        "properties": {
          "channel": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "image_id": {
            "title": "Image Id",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "osd_offset_us": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Osd Offset Us"
          },
          "overridden_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overridden By"
          },
          "residual_ms": {
            "title": "Residual Ms",
            "type": "number"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/ClockSegment"
            },
            "title": "Segments",
            "type": "array"
          }
        },
        "required": [
          "id",
          "image_id",
          "channel",
          "segments",
          "osd_offset_us",
          "confidence",
          "residual_ms",
          "method",
          "overridden_by"
        ],
        "title": "ClockModel",
        "type": "object"
      },
      "ClockOverrideRequest": {
        "additionalProperties": false,
        "properties": {
          "offset_us": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offset Us"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ClockOverrideRequest",
        "type": "object"
      },
      "ClockSegment": {
        "additionalProperties": false,
        "properties": {
          "from_device_us": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Device Us"
          },
          "offset_us": {
            "title": "Offset Us",
            "type": "integer"
          },
          "to_device_us": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Device Us"
          }
        },
        "required": [
          "from_device_us",
          "to_device_us",
          "offset_us"
        ],
        "title": "ClockSegment",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TimelineEvent": {
        "additionalProperties": false,
        "properties": {
          "channels": {
            "items": {
              "type": "integer"
            },
            "title": "Channels",
            "type": "array"
          },
          "end": {
            "title": "End",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "start": {
            "title": "Start",
            "type": "integer"
          }
        },
        "required": [
          "start",
          "end",
          "channels",
          "kind"
        ],
        "title": "TimelineEvent",
        "type": "object"
      },
      "TimelineMarker": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "log_event_id": {
            "title": "Log Event Id",
            "type": "string"
          },
          "ts_norm_us": {
            "title": "Ts Norm Us",
            "type": "integer"
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User"
          }
        },
        "required": [
          "ts_norm_us",
          "kind",
          "log_event_id",
          "user"
        ],
        "title": "TimelineMarker",
        "type": "object"
      },
      "TimelineResponse": {
        "additionalProperties": false,
        "properties": {
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ChannelTimeline"
            },
            "title": "Channels",
            "type": "array"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "markers": {
            "items": {
              "$ref": "#/components/schemas/TimelineMarker"
            },
            "title": "Markers",
            "type": "array"
          }
        },
        "required": [
          "channels",
          "markers",
          "events"
        ],
        "title": "TimelineResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Multi-vendor DVR/NVR forensic platform \u2014 API (docs/02-BACKEND.md).",
    "title": "Pramaan API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/cases/{cid}/clock-models": {
      "get": {
        "operationId": "list_clock_models_api_cases__cid__clock_models_get",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClockModel"
                  },
                  "title": "Response List Clock Models Api Cases  Cid  Clock Models Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Clock Models",
        "tags": [
          "timeline"
        ]
      }
    },
...
    "/api/cases/{cid}/timeline": {
      "get": {
        "operationId": "get_timeline_api_cases__cid__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "bucket_s",
            "required": false,
            "schema": {
              "default": 60,
              "title": "Bucket S",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Timeline",
        "tags": [
          "timeline"
        ]
      }
    },
...
    "/api/clock-models/{id}/override": {
      "post": {
        "operationId": "override_clock_model_api_clock_models__id__override_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClockOverrideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClockModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Override Clock Model",
        "tags": [
          "timeline"
        ]
      }
    },
...
  }
}